backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pydantic==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.11
better-exceptions==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography==43.0.0
uvicorn[standard]==0.32.0
